requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
google-generativeai>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pypdf>=3.17.0
streamlit>=1.24.0
requests>=2.31.0
python-multipart>=0.0.6
